changing the model of workout and get the app startable

+++ app/src/main/java/com/example/gymlog/data/local/room/entities/WorkoutResultEntity.kt
@@ -27,6 +27,6 @@ data class WorkoutResultEntity(
     val workTime: Int?,
     val sequenceInGymDay: Int,
     val position: Int,
-    val timeFromStart: Int,
+    val timeFromStart: Long,
     val workoutDateTime: String  // Обов'язкове поле
 )
